To the UI: -Added a newuser page that can create new admins, employees, or customers -Added a button to select creating a new user To the Server: -Added Username and password to both employee and customer
